third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-flexbox/position-absolute-013-expected.txt

This is a testharness.js-based test.
Found 432 tests; 288 PASS, 144 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS .flexbox 1
PASS .flexbox 2
PASS .flexbox 3
PASS .flexbox 4
PASS .flexbox 5
PASS .flexbox 6
PASS .flexbox 7
PASS .flexbox 8
PASS .flexbox 9
PASS .flexbox 10
PASS .flexbox 11
PASS .flexbox 12
PASS .flexbox 13
PASS .flexbox 14
PASS .flexbox 15
PASS .flexbox 16
PASS .flexbox 17
PASS .flexbox 18
PASS .flexbox 19
PASS .flexbox 20
PASS .flexbox 21
PASS .flexbox 22
PASS .flexbox 23
PASS .flexbox 24
PASS .flexbox 25
PASS .flexbox 26
PASS .flexbox 27
PASS .flexbox 28
PASS .flexbox 29
PASS .flexbox 30
PASS .flexbox 31
PASS .flexbox 32
PASS .flexbox 33
PASS .flexbox 34
PASS .flexbox 35
PASS .flexbox 36
PASS .flexbox 37
PASS .flexbox 38
PASS .flexbox 39
PASS .flexbox 40
PASS .flexbox 41
PASS .flexbox 42
PASS .flexbox 43
PASS .flexbox 44
PASS .flexbox 45
PASS .flexbox 46
PASS .flexbox 47
PASS .flexbox 48
PASS .flexbox 49
PASS .flexbox 50
PASS .flexbox 51
PASS .flexbox 52
PASS .flexbox 53
PASS .flexbox 54
PASS .flexbox 55
PASS .flexbox 56
PASS .flexbox 57
PASS .flexbox 58
PASS .flexbox 59
PASS .flexbox 60
PASS .flexbox 61
PASS .flexbox 62
PASS .flexbox 63
PASS .flexbox 64
PASS .flexbox 65
PASS .flexbox 66
PASS .flexbox 67
PASS .flexbox 68
PASS .flexbox 69
PASS .flexbox 70
PASS .flexbox 71
PASS .flexbox 72
PASS .flexbox 73
PASS .flexbox 74
PASS .flexbox 75
PASS .flexbox 76
PASS .flexbox 77
PASS .flexbox 78
PASS .flexbox 79
PASS .flexbox 80
PASS .flexbox 81
PASS .flexbox 82
PASS .flexbox 83
PASS .flexbox 84
PASS .flexbox 85
PASS .flexbox 86
PASS .flexbox 87
PASS .flexbox 88
PASS .flexbox 89
PASS .flexbox 90
PASS .flexbox 91
PASS .flexbox 92
PASS .flexbox 93
PASS .flexbox 94
PASS .flexbox 95
PASS .flexbox 96
PASS .flexbox 97
PASS .flexbox 98
PASS .flexbox 99
PASS .flexbox 100
PASS .flexbox 101
PASS .flexbox 102
PASS .flexbox 103
PASS .flexbox 104
PASS .flexbox 105
PASS .flexbox 106
PASS .flexbox 107
PASS .flexbox 108
PASS .flexbox 109
PASS .flexbox 110
PASS .flexbox 111
PASS .flexbox 112
PASS .flexbox 113
PASS .flexbox 114
PASS .flexbox 115
PASS .flexbox 116
PASS .flexbox 117
PASS .flexbox 118
PASS .flexbox 119
PASS .flexbox 120
PASS .flexbox 121
PASS .flexbox 122
PASS .flexbox 123
PASS .flexbox 124
PASS .flexbox 125
PASS .flexbox 126
PASS .flexbox 127
PASS .flexbox 128
PASS .flexbox 129
PASS .flexbox 130
PASS .flexbox 131
PASS .flexbox 132
PASS .flexbox 133
PASS .flexbox 134
PASS .flexbox 135
PASS .flexbox 136
PASS .flexbox 137
PASS .flexbox 138
PASS .flexbox 139
PASS .flexbox 140
PASS .flexbox 141
PASS .flexbox 142
PASS .flexbox 143
PASS .flexbox 144
FAIL .flexbox 145 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-start align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 146 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-start align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 147 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-start align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 148 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-end align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 149 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-end align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 150 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-end align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 151 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-start align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 152 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-start align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 153 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-start align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 154 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-end align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 155 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-end align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 156 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-end align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 157 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-start align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 158 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-start align-items-flex-end nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 159 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-start align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 160 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-end align-items-flex-start nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 161 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-end align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 162 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-end align-items-stretch nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 163 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-start align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 164 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-start align-items-flex-end nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 165 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-start align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 166 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-end align-items-flex-start nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 167 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-end align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 168 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-end align-items-stretch nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 169 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-start align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 170 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-start align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 171 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-start align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 172 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-end align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 173 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-end align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 174 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-end align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 175 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-start align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 176 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-start align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 177 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-start align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 178 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-end align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 179 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-end align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 180 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-end align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 181 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-start align-items-flex-start nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 182 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-start align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 183 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-start align-items-stretch nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 184 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-end align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 185 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-end align-items-flex-end nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 186 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-end align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 187 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-start align-items-flex-start nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 188 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-start align-items-flex-end nowrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 189 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-start align-items-stretch nowrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 190 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-end align-items-flex-start nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 191 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-end align-items-flex-end nowrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 192 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-end align-items-stretch nowrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 193 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-start align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 194 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-start align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 195 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-start align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 196 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-end align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 197 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-end align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 198 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-end align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 199 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-start align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 200 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-start align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 201 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-start align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 202 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-end align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 203 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-end align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 204 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-end align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 205 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-start align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 206 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-start align-items-flex-end wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 207 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-start align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 208 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-end align-items-flex-start wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 209 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-end align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 210 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-end align-items-stretch wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 211 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-start align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 212 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-start align-items-flex-end wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 213 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-start align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 214 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-end align-items-flex-start wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 215 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-end align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 216 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-end align-items-stretch wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 217 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-start align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 218 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-start align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 219 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-start align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 220 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-end align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 221 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-end align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 222 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-end align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 223 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-start align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 224 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-start align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 225 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-start align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 226 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-end align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 227 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-end align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 228 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-end align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 229 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-start align-items-flex-start wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 230 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-start align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 231 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-start align-items-stretch wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 232 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-end align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 233 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-end align-items-flex-end wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 234 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-end align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 235 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-start align-items-flex-start wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 236 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-start align-items-flex-end wrap"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 237 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-start align-items-stretch wrap"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 238 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-end align-items-flex-start wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 239 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-end align-items-flex-end wrap"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 240 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-end align-items-stretch wrap"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 241 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-start align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 242 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-start align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 243 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-start align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 244 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-end align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 245 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-end align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 246 assert_equals: 
<div class="flexbox vertical-rl ltr row justify-content-flex-end align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 247 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-start align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 248 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-start align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 249 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-start align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 250 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-end align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 251 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-end align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 252 assert_equals: 
<div class="flexbox vertical-rl rtl row justify-content-flex-end align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 253 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-start align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 254 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-start align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 255 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-start align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 256 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-end align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 257 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-end align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 258 assert_equals: 
<div class="flexbox vertical-rl ltr column justify-content-flex-end align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 259 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-start align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 260 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-start align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 261 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-start align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 262 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-end align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 263 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-end align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 264 assert_equals: 
<div class="flexbox vertical-rl rtl column justify-content-flex-end align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 265 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-start align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 266 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-start align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 267 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-start align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 268 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-end align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 269 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-end align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 270 assert_equals: 
<div class="flexbox vertical-rl ltr row-reverse justify-content-flex-end align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 271 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-start align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 272 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-start align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 273 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-start align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 274 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-end align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 275 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-end align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 276 assert_equals: 
<div class="flexbox vertical-rl rtl row-reverse justify-content-flex-end align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 277 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-start align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 278 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-start align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 279 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-start align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 280 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-end align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 281 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-end align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 282 assert_equals: 
<div class="flexbox vertical-rl ltr column-reverse justify-content-flex-end align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 283 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-start align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 284 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-start align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="50"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 285 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-start align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="0" data-offset-y="0"></div></div>
offsetLeft expected 0 but got 669
FAIL .flexbox 286 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-end align-items-flex-start wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 287 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-end align-items-flex-end wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="50"></div></div>
offsetLeft expected 50 but got 719
FAIL .flexbox 288 assert_equals: 
<div class="flexbox vertical-rl rtl column-reverse justify-content-flex-end align-items-stretch wrap-reverse"><div class="abspos" data-offset-x="50" data-offset-y="0"></div></div>
offsetLeft expected 50 but got 719
PASS .flexbox 289
PASS .flexbox 290
PASS .flexbox 291
PASS .flexbox 292
PASS .flexbox 293
PASS .flexbox 294
PASS .flexbox 295
PASS .flexbox 296
PASS .flexbox 297
PASS .flexbox 298
PASS .flexbox 299
PASS .flexbox 300
PASS .flexbox 301
PASS .flexbox 302
PASS .flexbox 303
PASS .flexbox 304
PASS .flexbox 305
PASS .flexbox 306
PASS .flexbox 307
PASS .flexbox 308
PASS .flexbox 309
PASS .flexbox 310
PASS .flexbox 311
PASS .flexbox 312
PASS .flexbox 313
PASS .flexbox 314
PASS .flexbox 315
PASS .flexbox 316
PASS .flexbox 317
PASS .flexbox 318
PASS .flexbox 319
PASS .flexbox 320
PASS .flexbox 321
PASS .flexbox 322
PASS .flexbox 323
PASS .flexbox 324
PASS .flexbox 325
PASS .flexbox 326
PASS .flexbox 327
PASS .flexbox 328
PASS .flexbox 329
PASS .flexbox 330
PASS .flexbox 331
PASS .flexbox 332
PASS .flexbox 333
PASS .flexbox 334
PASS .flexbox 335
PASS .flexbox 336
PASS .flexbox 337
PASS .flexbox 338
PASS .flexbox 339
PASS .flexbox 340
PASS .flexbox 341
PASS .flexbox 342
PASS .flexbox 343
PASS .flexbox 344
PASS .flexbox 345
PASS .flexbox 346
PASS .flexbox 347
PASS .flexbox 348
PASS .flexbox 349
PASS .flexbox 350
PASS .flexbox 351
PASS .flexbox 352
PASS .flexbox 353
PASS .flexbox 354
PASS .flexbox 355
PASS .flexbox 356
PASS .flexbox 357
PASS .flexbox 358
PASS .flexbox 359
PASS .flexbox 360
PASS .flexbox 361
PASS .flexbox 362
PASS .flexbox 363
PASS .flexbox 364
PASS .flexbox 365
PASS .flexbox 366
PASS .flexbox 367
PASS .flexbox 368
PASS .flexbox 369
PASS .flexbox 370
PASS .flexbox 371
PASS .flexbox 372
PASS .flexbox 373
PASS .flexbox 374
PASS .flexbox 375
PASS .flexbox 376
PASS .flexbox 377
PASS .flexbox 378
PASS .flexbox 379
PASS .flexbox 380
PASS .flexbox 381
PASS .flexbox 382
PASS .flexbox 383
PASS .flexbox 384
PASS .flexbox 385
PASS .flexbox 386
PASS .flexbox 387
PASS .flexbox 388
PASS .flexbox 389
PASS .flexbox 390
PASS .flexbox 391
PASS .flexbox 392
PASS .flexbox 393
PASS .flexbox 394
PASS .flexbox 395
PASS .flexbox 396
PASS .flexbox 397
PASS .flexbox 398
PASS .flexbox 399
PASS .flexbox 400
PASS .flexbox 401
PASS .flexbox 402
PASS .flexbox 403
PASS .flexbox 404
PASS .flexbox 405
PASS .flexbox 406
PASS .flexbox 407
PASS .flexbox 408
PASS .flexbox 409
PASS .flexbox 410
PASS .flexbox 411
PASS .flexbox 412
PASS .flexbox 413
PASS .flexbox 414
PASS .flexbox 415
PASS .flexbox 416
PASS .flexbox 417
PASS .flexbox 418
PASS .flexbox 419
PASS .flexbox 420
PASS .flexbox 421
PASS .flexbox 422
PASS .flexbox 423
PASS .flexbox 424
PASS .flexbox 425
PASS .flexbox 426
PASS .flexbox 427
PASS .flexbox 428
PASS .flexbox 429
PASS .flexbox 430
PASS .flexbox 431
PASS .flexbox 432
Harness: the test ran to completion.
